2-ai-agent-basics/pyproject.toml
[project]
name = "ai-agent-basics"
version = "0.1.0"
description = "Learn about agents directly hands on"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-genai>=1.12.1",
    "jupyter>=1.1.1",
    "mcp[cli]>=1.6.0",
    "numpy>=2.2.4",
    "ollama>=0.4.7",
]
